+++ Assets/Scripts/Save.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 [System.Serializable]
 public class Save
 {
-    public decimal savBalance;
+    public float savBalance;
 
-    public decimal savMultiplier;
+    public float savMultiplier;
 
-    public Dictionary<Rarity, Dictionary<string, decimal>> savTrashCollectCount;
+    public Dictionary<Rarity, Dictionary<string, float>> savTrashCollectCount;
     
     public Dictionary<WearableItem, string> savCustomCharacter;
